// apps/desktop/src/pdf/types.ts
/*
MIT License

Copyright (c) 2025 Bloodawn

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to do so, subject to the
following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.
*/

export interface PdfIndexRequest {
  type: "index";
  fileId: string;
  fileName: string;
  jumpId?: string | null;
  arrayBuffer: ArrayBuffer;
}

export interface PdfIndexProgress {
  type: "progress";
  fileId: string;
  currentPage: number;
  totalPages: number;
  percent: number;
}

export interface PdfIndexComplete {
  type: "complete";
  fileId: string;
  text: string;
  totalPages: number;
  emptyPages: number;
}

export interface PdfIndexError {
  type: "error";
  fileId: string;
  message: string;
}

export type PdfWorkerOutbound = PdfIndexProgress | PdfIndexComplete | PdfIndexError;

export type PdfWorkerInbound = PdfIndexRequest;